backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
python-multipart==0.0.6
pillow>=10.0.1
tensorflow>=2.13.0
numpy>=1.24.0
python-dotenv==1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymongo>=4.6.0
motor>=3.3.2
twilio>=8.10.0
reportlab>=4.0.7
pydantic[email]>=2.5.0
